Move flatten_list to to utils
